# Prework - *About Me*
Submitted by: **Mageto Nyakoni**

**About Me** is an app that... showcases a bit about me as a student and randomizes the background color of the app!

Time spent: **2** hours spent in total

## Required Features
Button Funcitonality to randomise the color
The following **required** functionality is completed:

- [x ] Users are see a screen with three labels and a button
- [x ] Tapping the button changes the screen color to a random color

## Video Walkthrough

Here is a reminder on how to embed Loom videos on GitHub. Feel free to remove this reminder once you upload your README.

[Guide]](https://www.youtube.com/watch?v=GA92eKlYio4) .

## App Brainstorming (Step 4)
- Levels (Social Media Workout Tracker):
    - I love the interactability with friends!
    - The progresssion system after every workout makes me feel accomplished
- Duolingo
    - I like the friendly reminders the app provides to logging in
    - love the cute little widget to remind me to renew my streak
    - the UI is very simple and aesthetically pleasing
- Maps
    - I like the interaction with the dynamic islands when possible
    - The use of teh always on display is cool, so you always know where you're going
    - the audio as well as visual cues are nice!

I would love to create a personal audiojournalling app to rreflect on books, media and serve as a journal for myself, as I prefer audiojournalling to
traditional written forms of journalling. maybe implement a streak feature and a notification as a reminder to login every day, as well as an award system to
incentivize reflection on a daily basis.
## Notes

Describe any challenges encountered while building the app:
I had issues seeing the color being randomized while running the program, as well as linking the button to the code, however playing around
for a while and closing xcode ultimatley lead to the results I needed

## License

    Copyright [2024] [Mageto Nyakoni]

    Licensed under the Apache License, Version 2.0 (the "License");
    you may not use this file except in compliance with the License.
    You may obtain a copy of the License at

        http://www.apache.org/licenses/LICENSE-2.0

    Unless required by applicable law or agreed to in writing, software
    distributed under the License is distributed on an "AS IS" BASIS,
    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
    See the License for the specific language governing permissions and
    limitations under the License.